+++ src/cspark/sdk/resources/_async/_services.py
@@ -21,6 +21,13 @@ class AsyncServices(AsyncApiResource):
     def compilation(self) -> 'AsyncCompilation':
         return AsyncCompilation(self.config, self._client)
 
+    async def exists(
+        self, uri: Union[None, str, UriParams] = None, *, folder: Optional[str] = None, service: Optional[str] = None
+    ):
+        response = await self.check_existence(uri or UriParams(folder, service))
+        data = response.data.get('response_data', {}) if isinstance(response.data, dict) else {}
+        return response.status == 200 and data.get('is_exists', False) is True
+
     async def create(
         self,
         name: str,
@@ -465,6 +472,13 @@ class AsyncServices(AsyncApiResource):
         url = Uri.of(base_url=self.config.base_url.value, version='api/v1', endpoint=endpoint)
 
         return await self.request(url, method='DELETE')
+
+    async def check_existence(
+        self, uri: Union[None, str, UriParams] = None, *, folder: Optional[str] = None, service: Optional[str] = None
+    ):
+        uri = Uri.validate(uri or UriParams(folder, service))
+        url = Uri.of(uri, base_url=self.config.base_url.full, endpoint='exists')
+        return await self.request(url)
 
     def __encode(
         self,